move permute_vector to utils
